pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "context-labeler"
version = "0.1.0"
description = ""
requires-python = ">=3.10"
dependencies = [
  "google-genai>=1.52.0",
]

[project.optional-dependencies]
dev = [
]